Move `CellGrid` from `typst-layout` to `typst-library`.
